# app/Http/Controllers/Api/UserController.php
<?php

namespace App\Http\Controllers\Api;

use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Http\JsonResponse;
use Illuminate\Support\Facades\Storage;
use App\Http\Controllers\Controller;
use Illuminate\Support\Facades\Hash;
use Illuminate\Support\Str;
use Illuminate\Support\Facades\Validator;

/**
 * @OA\Tag(
 *     name="Users",
 *     description="API Endpoints untuk manajemen user"
 * )
 */
class UserController extends Controller
{

    /**
     * @OA\Get(
     *     path="/myinfo",
     *     tags={"Users"},
     *     summary="Get logged-in user info",
     *     description="Mengambil detail informasi akun user yang sedang login (berdasarkan token).",
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Informasi user yang sedang login",
     *         @OA\JsonContent(
     *             @OA\Property(property="id_user", type="integer", example=1),
     *             @OA\Property(property="username", type="string", example="johndoe"),
     *             @OA\Property(property="name", type="string", example="John Doe"),
     *             @OA\Property(property="email", type="string", example="john@gmail.com"),
     *             @OA\Property(property="role", type="string", example="pembaca"),
     *             @OA\Property(property="membership", type="string", example="free"),
     *             @OA\Property(property="foto_profil", type="string", example="https://example.com/storage/profil/profile.jpg")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized - Token tidak valid atau belum login"
     *     )
     * )
     */
    public function myInfo(Request $request): JsonResponse
    {
        $user = $request->user(); // Dapatkan user dari token login (auth:sanctum)

        if (!$user) {
            return response()->json(['message' => 'Unauthorized'], 401);
        }

        return response()->json([
            'id_user' => $user->id_user,
            'username' => $user->username,
            'name' => $user->name,
            'email' => $user->email,
            'role' => $user->role,
            'membership' => $user->membership,
            'foto_profil' => $user->foto_profil
                ? asset('storage/' . $user->foto_profil)
                : null
        ]);
    }

    /**
     * @OA\Get(
     *     path="/users",
     *     tags={"Users"},
     *     summary="Get all users",
     *     description="Mengambil daftar semua user",
     *     @OA\Response(
     *         response=200,
     *         description="List semua user",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(
     *                 @OA\Property(property="id_user", type="integer", example=1),
     *                 @OA\Property(property="username", type="string", example="johndoe"),
     *                 @OA\Property(property="name", type="string", example="John Doe"),
     *                 @OA\Property(property="email", type="string", example="john@gmail.com"),
     *                 @OA\Property(property="role", type="string", example="pembaca"),
     *                 @OA\Property(property="membership", type="string", example="free"),
     *                 @OA\Property(property="foto_profil", type="string", example="profile.jpg")
     *             )
     *         )
     *     )
     * )
     */
    public function index(): JsonResponse
    {
        $users = User::all()->map(fn($user) => [
            'id_user' => $user->id_user,
            'username' => $user->username,
            'name' => $user->name,
            'email' => $user->email,
            'role' => $user->role,
            'membership' => $user->membership,
            'foto_profil' => $user->foto_profil
        ]);

        return response()->json($users);
    }

    /**
     * @OA\Get(
     *     path="/users/{id_user}",
     *     tags={"Users"},
     *     summary="Get single user",
     *     description="Mengambil detail user berdasarkan ID",
     *     @OA\Parameter(
     *         name="id_user",
     *         in="path",
     *         description="ID user",
     *         required=true,
     *         @OA\Schema(type="integer", example=1)
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Detail user",
     *         @OA\JsonContent(
     *             @OA\Property(property="id_user", type="integer", example=1),
     *             @OA\Property(property="username", type="string", example="johndoe"),
     *             @OA\Property(property="role", type="string", example="pembaca"),
     *             @OA\Property(property="membership", type="string", example="free"),
     *             @OA\Property(property="foto_profil", type="string", example="profile.jpg")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="User tidak ditemukan"
     *     )
     * )
     */
    public function show($id_user): JsonResponse
    {
        $user = User::find($id_user);
        if (!$user) {
            return response()->json(['error' => 'User tidak ditemukan'], 404);
        }

        return response()->json([
            'id_user' => $user->id_user,
            'username' => $user->username,
            'role' => $user->role,
            'membership' => $user->membership,
            'foto_profil' => $user->foto_profil
        ]);
    }


    /**
     * @OA\Post(
     *     path="/users/{id_user}",
     *     tags={"Users"},
     *     summary="Update user profile (self only)",
     *     description="Mengupdate profil user (username, name, email, password, foto). Hanya bisa dilakukan oleh user itu sendiri.",
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="id_user",
     *         in="path",
     *         description="ID user yang ingin diupdate (harus sama dengan user yang login)",
     *         required=true,
     *         @OA\Schema(type="integer", example=1)
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\MediaType(
     *             mediaType="multipart/form-data",
     *             @OA\Schema(
     *                 @OA\Property(property="username", type="string", example="johndoe"),
     *                 @OA\Property(property="name", type="string", example="John Doe"),
     *                 @OA\Property(property="email", type="string", example="john@gmail.com"),
     *                 @OA\Property(property="old_password", type="string", format="password", example="P@ssW0rd3"),
     *                 @OA\Property(property="new_password", type="string", format="password", example="newP@ssW0rd3"),
     *                 @OA\Property(property="foto_profil", type="string", format="binary")
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="User berhasil diupdate",
     *         @OA\JsonContent(
     *             @OA\Property(property="id_user", type="integer", example=1),
     *             @OA\Property(property="username", type="string", example="johndoe"),
     *             @OA\Property(property="name", type="string", example="John Doe"),
     *             @OA\Property(property="email", type="string", example="john@gmail.com"),
     *             @OA\Property(property="role", type="string", example="pembaca"),
     *             @OA\Property(property="membership", type="string", example="free"),
     *             @OA\Property(property="foto_profil", type="string", example="profil/profile.jpg")
     *         )
     *     ),
     *     @OA\Response(response=400, description="Password lama salah"),
     *     @OA\Response(response=401, description="Unauthenticated"),
     *     @OA\Response(response=403, description="Akses ditolak (bukan user sendiri)"),
     *     @OA\Response(response=404, description="User tidak ditemukan")
     * )
     */
    public function update(Request $request, $id_user): JsonResponse
    {
        $authUser = $request->user();

        // Pastikan sudah login
        if (!$authUser) {
            return response()->json(['message' => 'Unauthenticated'], 401);
        }

        // Cegah update user lain
        if ($authUser->id_user != $id_user) {
            return response()->json(['message' => 'Akses ditolak, Anda hanya dapat mengubah profil Anda sendiri'], 403);
        }

        // Temukan user
        $user = User::find($id_user);
        if (!$user) {
            return response()->json(['error' => 'User tidak ditemukan'], 404);
        }

        // Validasi input
        $validated = $request->validate([
            'username' => 'sometimes|string|max:255|unique:user,username,' . $id_user . ',id_user',
            'name' => 'sometimes|string|max:255',
            'email' => 'sometimes|string|email|max:255',
            'old_password' => 'sometimes|required_with:new_password|string',
            'new_password' => [
                'sometimes',
                'string',
                'min:8',
                'regex:/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$#!%*?&]+$/'
            ],
            'foto_profil' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048'
        ]);

        // Ganti password jika dikirim
        if ($request->filled('new_password')) {
            if (!Hash::check($request->old_password, $user->password)) {
                return response()->json(['error' => 'Password lama salah'], 400);
            }
            $user->password = Hash::make($request->new_password);
        }

        // Update data biasa
        $user->fill($request->only(['username', 'name', 'email']));

        // Upload foto baru jika ada
        if ($request->hasFile('foto_profil')) {
            if ($user->foto_profil && Storage::disk('public')->exists($user->foto_profil)) {
                Storage::disk('public')->delete($user->foto_profil);
            }

            $file = $request->file('foto_profil');
            $infofile = pathinfo($file->getClientOriginalName());
            $uuidgen = (string) Str::uuid(); // Returns a Uuid object
            $filename = time() . '_' . $uuidgen . '.' . $infofile['extension'];
            $path = $file->storeAs('profil', $filename, 'public');
            $user->foto_profil = $path;
        }

        $user->save();

        return response()->json([
            'id_user' => $user->id_user,
            'username' => $user->username,
            'name' => $user->name,
            'email' => $user->email,
            'role' => $user->role,
            'membership' => $user->membership,
            'foto_profil' => $user->foto_profil ? asset('storage/' . $user->foto_profil) : null
        ]);
    }



    /**
     * @OA\Delete(
     *     path="/users/{id_user}",
     *     tags={"Users"},
     *     summary="Delete user (admin only)",
     *     description="Menghapus user dari sistem. Hanya bisa dilakukan oleh admin.",
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="id_user",
     *         in="path",
     *         description="ID user yang ingin dihapus",
     *         required=true,
     *         @OA\Schema(type="integer", example=1)
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="User berhasil dihapus",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="User berhasil dihapus")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     ),
     *     @OA\Response(
     *         response=403,
     *         description="Akses ditolak (bukan admin)"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="User tidak ditemukan"
     *     )
     * )
     */
    public function destroy(Request $request, $id_user): JsonResponse
    {
        $authUser = $request->user();

        // Pastikan user sudah login
        if (!$authUser) {
            return response()->json(['message' => 'Unauthenticated'], 401);
        }

        // Hanya admin yang bisa hapus user
        if ($authUser->role !== 'admin') {
            return response()->json(['message' => 'Akses ditolak, hanya admin yang dapat menghapus user'], 403);
        }

        // Temukan user berdasarkan ID
        $user = User::find($id_user);
        if (!$user) {
            return response()->json(['error' => 'User tidak ditemukan'], 404);
        }

        // Jangan izinkan admin menghapus dirinya sendiri (opsional, bisa dihapus jika tidak diinginkan)
        if ($authUser->id_user == $user->id_user) {
            return response()->json(['message' => 'Admin tidak dapat menghapus akun sendiri'], 403);
        }

        $user->delete();

        return response()->json(['message' => 'User berhasil dihapus']);
    }

    /**
     * @OA\Post(
     * path="/users/{id_user}/role",
     * summary="Memperbarui role user",
     * description="Hanya dapat diakses oleh admin untuk mengubah role pengguna lain atau dirinya sendiri.",
     * tags={"Users"},
     * security={{"bearerAuth":{}}},
     * @OA\Parameter(
     * name="id_user",
     * in="path",
     * description="ID dari user yang akan diupdate",
     * required=true,
     * @OA\Schema(type="integer")
     * ),
     * @OA\RequestBody(
     * required=true,
     * @OA\JsonContent(
     * required={"role"},
     * @OA\Property(property="role", type="string", enum={"admin", "user", "editor"}, example="admin")
     * )
     * ),
     * @OA\Response(
     * response=200,
     * description="Role berhasil diperbarui",
     * @OA\JsonContent(
     * @OA\Property(property="message", type="string", example="Role user berhasil diperbarui"),
     * @OA\Property(property="data", type="object")
     * )
     * ),
     * @OA\Response(
     * response=401,
     * description="Unauthenticated",
     * @OA\JsonContent(@OA\Property(property="message", type="string", example="Unauthenticated"))
     * ),
     * @OA\Response(
     * response=403,
     * description="Akses ditolak (Bukan Admin atau melarang ganti role sendiri)",
     * @OA\JsonContent(@OA\Property(property="message", type="string", example="Akses ditolak, hanya admin yang dapat mengubah role"))
     * ),
     * @OA\Response(
     * response=404,
     * description="User tidak ditemukan",
     * @OA\JsonContent(@OA\Property(property="error", type="string", example="User tidak ditemukan"))
     * ),
     * @OA\Response(
     * response=422,
     * description="Validasi gagal",
     * @OA\JsonContent(@OA\Property(property="errors", type="object"))
     * )
     * )
     */
    public function updateRole(Request $request, $id_user): JsonResponse
    {
        $authUser = $request->user();

        // 1. Pastikan user sudah login
        if (!$authUser) {
            return response()->json(['message' => 'Unauthenticated'], 401);
        }

        // 2. Hanya admin yang bisa mengubah role
        if ($authUser->role !== 'admin') {
            return response()->json(['message' => 'Akses ditolak, hanya admin yang dapat mengubah role'], 403);
        }

        // 3. Validasi input role
        $validator = Validator::make($request->all(), [
            'role' => 'required|string|in:admin,pembaca,penulis', // Sesuaikan dengan daftar role di aplikasi Anda
        ]);

        if ($validator->fails()) {
            return response()->json(['errors' => $validator->errors()], 422);
        }

        // 4. Temukan user berdasarkan ID
        $user = User::find($id_user);
        if (!$user) {
            return response()->json(['error' => 'User tidak ditemukan'], 404);
        }

        // 5. Opsional: Mencegah admin mengubah role-nya sendiri menjadi non-admin
        // Agar tidak terjadi "lockout" (kehilangan akses admin secara tidak sengaja)
        if ($authUser->id_user == $user->id_user && $request->role !== 'admin') {
            return response()->json(['message' => 'Admins cannot change their own role to non-admin.'], 403);
        } else if ($request->role === 'admin') {
            return response()->json(['message' => 'You cannot change another user to an admin for security reasons through the admin panel.'], 403);
        }

        // 6. Update role
        $user->role = $request->role;
        $user->save();

        return response()->json([
            'message' => 'Role user berhasil diperbarui',
            'data' => [
                'id_user' => $user->id_user,
                'name' => $user->name,
                'role' => $user->role
            ]
        ]);
    }
}