# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Berita Documentation",
        "description": "Dokumentasi API untuk sistem manajemen berita",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/myinfo": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get logged-in user info",
                "description": "Mengambil detail informasi akun user yang sedang login (berdasarkan token).",
                "operationId": "1dc02b93a57db8d902e3867dc42787c2",
                "responses": {
                    "200": {
                        "description": "Informasi user yang sedang login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_user": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@gmail.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "pembaca"
                                        },
                                        "membership": {
                                            "type": "string",
                                            "example": "free"
                                        },
                                        "foto_profil": {
                                            "type": "string",
                                            "example": "https://example.com/storage/profil/profile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid atau belum login"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/users/{id_user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get single user",
                "description": "Mengambil detail user berdasarkan ID",
                "operationId": "163dadaa4e56547be21e2297b0992bf6",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_user": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "pembaca"
                                        },
                                        "membership": {
                                            "type": "string",
                                            "example": "free"
                                        },
                                        "foto_profil": {
                                            "type": "string",
                                            "example": "profile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user profile (self only)",
                "description": "Mengupdate profil user (username, name, email, password, foto). Hanya bisa dilakukan oleh user itu sendiri.",
                "operationId": "d21cabb937ce4befce14e5d7f6c4fb8d",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user yang ingin diupdate (harus sama dengan user yang login)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@gmail.com"
                                    },
                                    "old_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "P@ssW0rd3"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newP@ssW0rd3"
                                    },
                                    "foto_profil": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_user": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@gmail.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "pembaca"
                                        },
                                        "membership": {
                                            "type": "string",
                                            "example": "free"
                                        },
                                        "foto_profil": {
                                            "type": "string",
                                            "example": "profil/profile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Password lama salah"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Akses ditolak (bukan user sendiri)"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user (admin only)",
                "description": "Menghapus user dari sistem. Hanya bisa dilakukan oleh admin.",
                "operationId": "26b1d3f0d3f3d223b97c69f73802c588",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Akses ditolak (bukan admin)"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT"
            }
        }
    },
...
}